requirements.txt
cached-property==1.5.2
certifi==2021.5.30
cycler==0.11.0
dataclasses
h5py==3.1.0
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.6.0.66
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
zipp==3.6.0
